/deleteroom: Easily delete personal rooms
To avoid deleting the wrong room by accident, it is required to type out the room name.

This PR removes that requirement for personal rooms. It also addresses the following issues with the permission check:

(1) The auth check looks at ``room.isPersonal`` instead of ``targetRoom.isPersonal``.

This should allow any room owner to delete their room if they do it from a groupchat, even if they do not have the makeroom permission.

(2) The room is searched for before permissions are checked. That creates an information leak. Any user can do the following:
- Use on an existing room: /deleteroom - Access denied.
- Use on a non-existing room: The room 'yxxyxyxyxxyy' doesn't exist.